+++ src/main/java/blackjack/domain/participant/PlayerFactory.java
@@ -2,7 +2,6 @@ package blackjack.domain.participant;
 
 import java.util.ArrayList;
 import java.util.List;
-import blackjack.domain.common.Money;
 
 public class PlayerFactory {
