+++ src/test/java/blackjack/domain/judgement/JudgementResultTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.judgement;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.common.Money;
+import blackjack.domain.participant.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
