+++ src/test/java/blackjack/domain/participant/MoneyTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.common;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
